Refactor clase User, añadido enum Roles
