+++ src/auth/role/role.enum.ts
@@ -1,0 +1,4 @@
+export enum Role {
+  USER = "user",
+  ADMIN = "admin"
+}
