+++ src/shared/typeorm/typeorm.config.ts
@@ -1,0 +1,12 @@
+import { User } from "../../auth/user/user.model";
+import { DataSource } from "typeorm";
+
+export default new DataSource({
+  type: "postgres",
+  host: process.env.DATABASE_HOST,
+  port: 5434,
+  database: "rct",
+  username: "postgres",
+  password: "123",
+  entities: [User],
+})
